LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/skip-outbound-vt-before-reveal-expected.txt


FAIL when navigating away before revealing, never start a view transition assert_equals: expected "did reveal new page without transition" but got "did reveal old page"
